+++ Changes.md
@@ -2,3 +2,4 @@ Version Information
 ===================
 Version 2.9.0.1
   1. Conversion from ra_accordion block.
+  2. Ensure that you get a down arrow when there is a sub-menu item that is current.